+++ tests/testAccess.js
@@ -486,4 +486,91 @@ describe('Access control', function() {
 
     /**********************************************************************/
 
+    describe('16. delete PrairieSchedule course link', function() {
+        it('should succeed', function(callback) {
+            sqldb.query(sql.delete_ps_course_link, [], function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should enable access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 200, callback);
+        });
+    });
+
+    describe('17. delete all reservations', function() {
+        it('should succeed', function(callback) {
+            sqldb.query(sql.delete_all_reservations, [], function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should enable access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 200, callback);
+        });
+    });
+
+    describe('18. delete all access rules', function() {
+        it('should succeed', function(callback) {
+            sqldb.query(sql.delete_access_rules, [], function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should block access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 500, callback);
+        });
+    });
+
+    describe('19. insert exam-linked access rule', function() {
+        it('should succeed', function(callback) {
+            sqldb.query(sql.insert_ps_exam_access_rule, {assessment_id}, function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should block access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 500, callback);
+        });
+    });
+
+    describe('20. insert PrairieSchedule reservation', function() {
+        it('should succeed', function(callback) {
+            var params = {user_id: user.user_id};
+            sqldb.query(sql.insert_ps_reservation, params, function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should block access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 500, callback);
+        });
+        it('should block access to the assessment_instance before the reservation', function(callback) {
+            getAssessmentInstance(cookiesStudentExamBeforeReservation(), 500, callback);
+        });
+        it('should block access to the assessment_instance after the reservation', function(callback) {
+            getAssessmentInstance(cookiesStudentExamAfterReservation(), 500, callback);
+        });
+    });
+
+    describe('21. check in PrairieSchedule reservation', function() {
+        it('should succeed', function(callback) {
+            sqldb.query(sql.update_ps_reservation_to_checked_in, [], function(err, _result) {
+                if (ERR(err, callback)) return;
+                callback(null);
+            });
+        });
+        it('should enable access to the assessment_instance', function(callback) {
+            getAssessmentInstance(cookiesStudentExam(), 200, callback);
+        });
+        it('should block access to the assessment_instance before the reservation', function(callback) {
+            getAssessmentInstance(cookiesStudentExamBeforeReservation(), 500, callback);
+        });
+        it('should block access to the assessment_instance after the reservation', function(callback) {
+            getAssessmentInstance(cookiesStudentExamAfterReservation(), 500, callback);
+        });
+    });
+
+    /**********************************************************************/
+
 });