+++ ChangeLog.md
@@ -105,6 +105,8 @@
 
   * Add submission info modal with external grading stats (Nathan Walters).
 
+  * Add per-assessment PrairieSchedule exam linking (Matt West).
+
   * Fix broken file upload element (Nathan Walters).
 
   * Fix broken popover and improve assessment label styles (Nathan Walters).